// client/src/app/wiki/constants/wiki-format.enum.ts
export enum WikiFormat {
    Heading1 = 'heading1',
    Heading2 = 'heading2',
    Heading3 = 'heading3',
    Bold = 'bold',
    Italic = 'italic',
    Strikethrough = 'strikethrough',
    Code = 'code',
    Link = 'link',
    WikiLink = 'wikiLink',
    BulletList = 'bulletList',
    OrderedList = 'orderedList',
    TaskList = 'taskList',
    Quote = 'quote',
    CodeBlock = 'codeBlock'
}